OTHER: Cleanup NowPlaying.
Removed the idle timeout since it never worked as intended and
fixing it in a platform independent way was kind of impossible.

Fullscreen is now activated via menu and playback accelerators
are now working in fullscreen mode. A bit of a hack since I
couldn't figure out how to propagate the keyevent to the
main window accelerator group. Should be fixed some day.