unify playlist entry creation and population
Prior to this commit, Abraca first creates cells with "insert" or "add"
and then separately populates them using "set".

Certain accessibility services intercept the insertion before anything
was stored, examining the cell content via AbracaPlaylistMode.get_value.
This triggers requesting song information from xmms2, if needed,
resulting in an erroneous request of media id 0, which leads to
seemingly empty entries in the playlist.

Accessibility tools that trigger this behaviour are for example screen
magnifiers or even gnome-shell 3.12 itself, as long as it is not built
including the patch to initialize at-spi on demand
<https://bugzilla.gnome.org/show_bug.cgi?id=730118>.

This commit unifies insertion and population to a single call to
"insert_with_values", making interception impossible in between.