OTHER: Fix shutdown crash.
Previously we listened to columns_changed event to track
what columns to show in the FilterView. I think the reason
for this was to keep dynamic_columns up to date with the
current column order. This however had a side effect of
GTK emitting this event during shutdown when destroying
the columns, and while doing so our updating of the model
caused a crash.

Instead we now set the new dynamic columns via the remove
event, and inspect the available columns directly when getting
the configuration to save (columns, and their order).