OTHER: Fix bug when editing columns.
The reference to FilterModel used in FilterEditor
was pointing to the instance that existed at the time
when the dialog was opened. When selecting a new column,
or deselecting an existing one, the store is replaced,
and thus the next action must refer to the new version.

The fix was to simply access the FilterView model member
directly from the FilterEditor event handler.