add expected VALAFLAGS
The configure function of the Vala waftool is explicitly skipped to
check the Vala version manually directly afterwards. This however also
deactivates the automatic setting of the correct VALAFLAGS through this
configure function, which are expected to be set correctly by most
components of the Vala waftool. Specifying the flag manually restores
compilation.